Growing babies Signed-off-by: Windwaker <wjcrouse913@aim.com>

+++ src/main/java/org/spout/vanilla/protocol/event/entity/player/PlayerHealthEvent.java
@@ -58,7 +58,7 @@ public class PlayerHealthEvent extends ProtocolEvent {
 		return saturation;
 	}
 
-	public void setFoodSaturation(float saturation) {
+	public void setFoodSaturation(short saturation) {
 		this.saturation = saturation;
 	}
 }
